requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.68.1
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.0
certifi==2025.8.3
charset-normalizer==3.4.3
deep-translator==1.11.4
distro==1.9.0
docstring_parser==0.17.0
filetype==1.2.0
frozenlist==1.7.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
multidict==6.6.4
pillow==11.3.0
pip_system_certs==5.2
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-magic==0.4.27
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
